+++ cmd/influxd/launcher/launcher.go
@@ -57,6 +57,8 @@ import (
 	"github.com/influxdata/influxdb/v2/session"
 	"github.com/influxdata/influxdb/v2/snowflake"
 	"github.com/influxdata/influxdb/v2/source"
+	"github.com/influxdata/influxdb/v2/sqlite"
+	sqliteMigrations "github.com/influxdata/influxdb/v2/sqlite/migrations"
 	"github.com/influxdata/influxdb/v2/storage"
 	storageflux "github.com/influxdata/influxdb/v2/storage/flux"
 	"github.com/influxdata/influxdb/v2/storage/readservice"
@@ -88,7 +90,9 @@ import (
 )
 
 const (
-	// BoltStore stores all REST resources in boltdb.
+	// DiskStore stores all REST resources to disk in boltdb and sqlite.
+	DiskStore = "disk"
+	// BoltStore also stores all REST resources to disk in boltdb and sqlite. Kept for backwards-compatibility.
 	BoltStore = "bolt"
 	// MemoryStore stores all REST resources in memory (useful for testing).
 	MemoryStore = "memory"
@@ -110,6 +114,7 @@ type Launcher struct {
 	boltClient *bolt.Client
 	kvStore    kv.SchemaStore
 	kvService  *kv.Service
+	sqlStore   *sqlite.SqlStore
 
 	// storage engine
 	engine Engine
@@ -182,6 +187,12 @@ func (m *Launcher) Shutdown(ctx context.Context) error {
 	m.log.Info("Stopping", zap.String("service", "bolt"))
 	if err := m.boltClient.Close(); err != nil {
 		m.log.Error("Failed closing bolt", zap.Error(err))
+		errs = append(errs, err.Error())
+	}
+
+	m.log.Info("Stopping", zap.String("service", "sqlite"))
+	if err := m.sqlStore.Close(); err != nil {
+		m.log.Error("Failed closing sqlite", zap.Error(err))
 		errs = append(errs, err.Error())
 	}
 
@@ -270,28 +281,51 @@ func (m *Launcher) run(ctx context.Context, opts *InfluxdOpts) (err error) {
 	var flushers flushers
 	switch opts.StoreType {
 	case BoltStore:
-		store := bolt.NewKVStore(m.log.With(zap.String("service", "kvstore-bolt")), opts.BoltPath)
-		store.WithDB(m.boltClient.DB())
-		m.kvStore = store
+		m.log.Warn("Using --store=bolt is deprecated. Use --store=disk instead.")
+		fallthrough
+	case DiskStore:
+		kvStore := bolt.NewKVStore(m.log.With(zap.String("service", "kvstore-bolt")), opts.BoltPath)
+		kvStore.WithDB(m.boltClient.DB())
+		m.kvStore = kvStore
+
+		// If a sqlite-path is not specified, store sqlite db in the same directory as bolt with the default filename.
+		if opts.SqLitePath == "" {
+			opts.SqLitePath = filepath.Dir(opts.BoltPath) + "/" + sqlite.DefaultFilename
+		}
+		sqlStore, err := sqlite.NewSqlStore(opts.SqLitePath, m.log.With(zap.String("service", "sqlite")))
+		if err != nil {
+			m.log.Error("Failed opening sqlite store", zap.Error(err))
+			return err
+		}
+		m.sqlStore = sqlStore
+
 		if opts.Testing {
-			flushers = append(flushers, store)
+			flushers = append(flushers, kvStore, sqlStore)
 		}
 
 	case MemoryStore:
-		store := inmem.NewKVStore()
-		m.kvStore = store
+		kvStore := inmem.NewKVStore()
+		m.kvStore = kvStore
+
+		sqlStore, err := sqlite.NewSqlStore(sqlite.InmemPath, m.log.With(zap.String("service", "sqlite")))
+		if err != nil {
+			m.log.Error("Failed opening sqlite store", zap.Error(err))
+			return err
+		}
+		m.sqlStore = sqlStore
+
 		if opts.Testing {
-			flushers = append(flushers, store)
+			flushers = append(flushers, kvStore, sqlStore)
 		}
 
 	default:
-		err := fmt.Errorf("unknown store type %s; expected bolt or memory", opts.StoreType)
-		m.log.Error("Failed opening bolt", zap.Error(err))
+		err := fmt.Errorf("unknown store type %s; expected disk or memory", opts.StoreType)
+		m.log.Error("Failed opening metadata store", zap.Error(err))
 		return err
 	}
 
-	migrator, err := migration.NewMigrator(
-		m.log.With(zap.String("service", "migrations")),
+	boltMigrator, err := migration.NewMigrator(
+		m.log.With(zap.String("service", "bolt migrations")),
 		m.kvStore,
 		all.Migrations[:]...,
 	)
@@ -300,9 +334,17 @@ func (m *Launcher) run(ctx context.Context, opts *InfluxdOpts) (err error) {
 		return err
 	}
 
-	// apply migrations to metadata store
-	if err := migrator.Up(ctx); err != nil {
-		m.log.Error("Failed to apply migrations", zap.Error(err))
+	// apply migrations to the bolt metadata store
+	if err := boltMigrator.Up(ctx); err != nil {
+		m.log.Error("Failed to apply bolt migrations", zap.Error(err))
+		return err
+	}
+
+	sqliteMigrator := sqlite.NewMigrator(m.sqlStore, m.log.With(zap.String("service", "sqlite migrations")))
+
+	// apply migrations to the sqlite metadata store
+	if err := sqliteMigrator.Up(ctx, &sqliteMigrations.All{}); err != nil {
+		m.log.Error("Failed to apply sqlite migrations", zap.Error(err))
 		return err
 	}